# Conflicts:
#	app/src/main/java/com/example/yomikaze_app_kotlin/Presentation/Navigation/NavigationGraph.kt
#	app/src/main/java/com/example/yomikaze_app_kotlin/Presentation/Screens/Authentication/ForgotPassword/ForgotPasswordView.kt
#	app/src/main/java/com/example/yomikaze_app_kotlin/Presentation/Screens/Authentication/Login/LoginView.kt
#	app/src/main/java/com/example/yomikaze_app_kotlin/Presentation/Screens/Profile/EditProfile/EditProfileView.kt
#	app/src/main/java/com/example/yomikaze_app_kotlin/Presentation/Screens/Profile/ProfileView.kt

+++ app/src/main/java/com/example/yomikaze_app_kotlin/Presentation/Components/ComicCard/RankingComicCard/RankingComicCard.kt
@@ -20,7 +20,7 @@ import com.example.yomikaze_app_kotlin.Presentation.Components.ComicCard.ShareCo
 
 @Composable
 fun RankingComicCard(
-    comicId: Int,
+    comicId: Long,
     rankingNumber: Int,
     image: String,
     comicName: String,
